move path meta key to a separate module
